+++ sl-map-cli/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 0.1.1
 
-Print PPS HUD config string when generating a map
+Print PPS HUD config string and size recommendations when generating a map
+
+Fix max zoom level calculations, previously it used the minimum zoom level out
+of the one calculated for x and y axes but lower zoom levels are higher detail
+so it needed to be the maximum.
 
 ## 0.1.0